Revert "unmount Slcsp controller (#2816)"
This reverts commit 627183038580e73bfdf8e40ab0df8b4249850d2e.
