+++ config/routes.rb
@@ -448,6 +448,16 @@ Rails.application.routes.draw do
   end
   resources :translations
 
+  namespace :api, :defaults => {:format => 'xml'} do
+    namespace :v1 do
+      resources :slcsp, :only => []  do
+        collection do
+          post :plan
+        end
+      end
+    end
+  end
+
   ############################# TO DELETE BELOW ##############################
 
   # FIXME: Do this properly later
@@ -599,16 +609,6 @@ Rails.application.routes.draw do
           get :resume_resident_enrollment, on: :collection
           get :ridp_bypass, on: :collection
           get :find_sep, on: :collection
-        end
-      end
-
-      namespace :api, :defaults => {:format => 'xml'} do
-        namespace :v1 do
-          resources :slcsp, :only => []  do
-            collection do
-              post :plan
-            end
-          end
         end
       end
     end
