+++ spec/controllers/api/v1/slcsp_controller_spec.rb
@@ -2,7 +2,7 @@
 
 require 'rails_helper'
 
-describe IndividualMarket::Api::V1::SlcspController, skip: true do
+describe Api::V1::SlcspController do
 
   let(:request_xml) { File.read(Rails.root.join("spec", "test_data", "slcsp_payloads", "request.xml")) }
   let(:response_xml) { File.read(Rails.root.join("spec", "test_data", "slcsp_payloads", "response.xml")) }
@@ -11,7 +11,7 @@ describe IndividualMarket::Api::V1::SlcspController, skip: true do
   context "valid request" do
     it 'returns https status 200' do
       allow(HappyMapper).to receive(:parse).with(anything).and_return(HappyMapper.parse(request_xml))
-      allow_any_instance_of(IndividualMarket::Api::V1::SlcspController).to receive(:find_slcsp).with(anything).and_return(plan)
+      allow_any_instance_of(Api::V1::SlcspController).to receive(:find_slcsp).with(anything).and_return(plan)
       allow_any_instance_of(ActionController::Rendering).to receive(:render).and_return(response_xml)
 
       post :plan, :format => "xml"
@@ -23,10 +23,11 @@ describe IndividualMarket::Api::V1::SlcspController, skip: true do
   context "invalid request" do
     it 'returns https status 422' do
       allow(HappyMapper).to receive(:parse).with(anything).and_raise(Exception.new)
-      allow_any_instance_of(IndividualMarket::Api::V1::SlcspController).to receive(:find_slcsp).with(anything).and_return(plan)
+      allow_any_instance_of(Api::V1::SlcspController).to receive(:find_slcsp).with(anything).and_return(plan)
 
       post :plan, params: { format: :xml }
       expect(response.status).to eq(422)
     end
   end
+
 end
